more int fixes

+++ metadata/client.py
@@ -78,9 +78,9 @@ class PMClient(object):
         """Delete the object of type from query_hash."""
         ret = requests.delete(
             '{0}/{1}'.format(self.url, cls_type), params=query_hash, allow_redirects=True)
-        if ret.status_code / 100 == 2 or ret.status_code / 100 == 4:
+        if int(ret.status_code / 100) == 2 or int(ret.status_code / 100) == 4:
             return True
-        elif ret.status_code / 100 == 5:
+        elif int(ret.status_code / 100) == 5:
             raise PMClientError('Internal Server Error ({0}) {1}'.format(
                 ret.status_code, ret.content))
         else:
